// src/modules/area-estrategia-desarrollo-comercial/analisis_empresas/application/dto/create-analisis-empresa.dto.ts
export class CreateAnalisisEmpresaDto {
  ruc?: string;
  nombre?: string;
  nombre_empresa?: string;
  nombreEmpresa?: string;
  correo?: string;
  telefono_fijo?: string;
  telefonoFijo?: string;
  celular?: string;
  departamento?: string;
  distrito?: string;
  direccion?: string;
  sector?: string;
  estado?: string;
  descripcion?: string;
}